+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 otj-server
 ==========
 
-Future
+5.2.5
 -----
+* Fixes to work with new otj-conservedheader changes for overriding headers
+* Tests for overriding and conserving headers.
 * POM 271
 * Add Dmitry's test to verify HTTP POST with 404 reacts correctly.
+* Patch an issue in conserved headers.
 
 5.2.4
 -----